requirements/requirements_model.txt
pyyaml
mxnet-mkl
numpy
gluonnlp
scipy
transformers
faiss-gpu
sentencepiece
adjustText
datasets
watermark
scikit-learn
matplotlib